requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
confusable-homoglyphs==3.2.0
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
django-heroku==0.3.1
django-modeltranslation==0.16.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.2
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.3
whitenoise==5.2.0
wincertstore==0.2
